// app/src/generated/models/Ben_ltkcardseriesesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Ben_ltkcardseriesesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Ben_ltkcardseriesesstatecode = keyof typeof Ben_ltkcardseriesesstatecode;
export const Ben_ltkcardseriesesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Ben_ltkcardseriesesstatuscode = keyof typeof Ben_ltkcardseriesesstatuscode;

export interface Ben_ltkcardseriesesBase {
  ben_boardid: string;
  ben_cardid: string;
  ben_date?: string;
  ben_ltkcardseriesid: string;
  ben_name?: string;
  ben_serieskey: string;
  ben_shift: string;
  ben_value?: string;
  ben_valuejson?: string;
  ben_who?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Ben_ltkcardseriesesstatecode;
  statuscode?: Ben_ltkcardseriesesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Ben_ltkcardserieses extends Ben_ltkcardseriesesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}